C++ Stack Implementation with Array.cpp
The first code implements a stack using an array and provides basic stack operations such as push, pop, isEmpty, and display. It measures execution time and is limited to a maximum size of 100 elements.

The second code defines a stack using a vector, simplifying stack management by utilizing the dynamic size feature of vectors. It offers the same stack operations and output messages. This implementation is more efficient and flexible but doesn't measure execution time. Both codes demonstrate the fundamental concepts of a stack data structure in C++.